Create Problem9.sql
(User-defined Exception) Create a user-defined exception by the name of exp_check. Select the 
empname and joined date of all employees into a cursor from the emp table as shown in problem
Your program should calculate the experience of all the employees in years, and insert the 
empname and experience of each employee into tempp table. Create the table beforehand. 
If any employee has experience less than 2 years, the program should be aborted with a suitable 
message. Raise the user-defined exception exp_check to achieve this. Display the results on the 
screen using dbms_output.put_line.