Create Problem10.sql
(Function) Write a PL/SQL function to take three parameters, the sides of a triangle. The sides of 
the triangle should be accepted by the user. The function should return a Boolean value - true
if the triangle is valid, false otherwise. A triangle is valid if the length of each side is less than the 
sum of the lengths of the other two sides. Check if the dimensions entered by the user can form a 
valid triangle. 
Display the results on the screen using dbms_output.put_line.